@@ -291,12 +291,14 @@ public class JsonRpcProtocol extends SequentialCommProtocol implements HttpUtils
 		}
 	}
 
+	@Override
 	public void send( OutputStream ostream, CommMessage message, InputStream istream )
 		throws IOException {
 		setChannelInterface();
 		HttpUtils.send( ostream, message, istream, inInputPort, channel(), this );
 	}
 
+	@Override
 	public CommMessage recv_internal( InputStream istream, OutputStream ostream )
 		throws IOException {
 		if( checkStringParameter( Parameters.TRANSPORT, LSP ) ) {
@@ -323,12 +325,9 @@ public class JsonRpcProtocol extends SequentialCommProtocol implements HttpUtils
 		} else {
 			HttpParser parser = new HttpParser( istream );
 			HttpMessage message = parser.parse();
-			String charset = HttpUtils.getCharset( null, message );
+			String charset = HttpUtils.getResponseCharset( message );
 			HttpUtils.recv_checkForChannelClosing( message, channel() );
 
-			if( message.isError() ) {
-				throw new IOException( "HTTP error: " + new String( message.content(), charset ) );
-			}
 			if( inInputPort && message.type() != HttpMessage.Type.POST ) {
 				throw new UnsupportedMethodException( "Only HTTP method POST allowed", Method.POST );
 			}
@@ -401,6 +400,7 @@ public class JsonRpcProtocol extends SequentialCommProtocol implements HttpUtils
 		channelInterface = channel().parentPort().getInterface();
 	}
 
+	@Override
 	public CommMessage recv( InputStream istream, OutputStream ostream )
 		throws IOException {
 		setChannelInterface();
