@@ -53,6 +53,7 @@ public class XmlRpcProtocolFactory extends CommProtocolFactory {
 		transformer.setOutputProperty( OutputKeys.OMIT_XML_DECLARATION, "yes" );
 	}
 
+	@Override
 	public CommProtocol createInputProtocol( VariablePath configurationPath, URI location )
 		throws IOException {
 		return new XmlRpcProtocol(
@@ -60,11 +61,11 @@ public class XmlRpcProtocolFactory extends CommProtocolFactory {
 			location,
 			true,
 			transformer,
-			docBuilderFactory,
 			docBuilder,
 			commCore().interpreter() );
 	}
 
+	@Override
 	public CommProtocol createOutputProtocol( VariablePath configurationPath, URI location )
 		throws IOException {
 		return new XmlRpcProtocol(
@@ -72,7 +73,6 @@ public class XmlRpcProtocolFactory extends CommProtocolFactory {
 			location,
 			false,
 			transformer,
-			docBuilderFactory,
 			docBuilder,
 			commCore().interpreter() );
 	}
